backend/requirements.txt
# Core Framework
fastapi>=0.104.0,<0.110.0
uvicorn[standard]>=0.24.0,<0.30.0
gunicorn>=21.2.0,<22.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0

# Data Validation
pydantic>=2.0.0,<3.0.0

# Environment Variables
python-dotenv>=1.0.0,<2.0.0

# HTTP Client (para OneSignal API y DeepSeek API)
requests>=2.31.0,<3.0.0
httpx>=0.25.0,<0.28.0

# Authentication
google-auth>=2.23.0,<3.0.0
PyJWT>=2.8.0,<3.0.0

# File Upload Support
python-multipart>=0.0.6,<0.1.0
